Added enum AudioChannelLayout containing entries for all valid channel layouts (i.e. mono, stereo, quad, and 5.1).

+++ sources/SDL2Sharp/AudioDevice.cs
@@ -139,7 +139,7 @@ namespace SDL2Sharp
             {
                 freq = spec.Frequency,
                 format = (ushort)spec.Format,
-                channels = spec.Channels,
+                channels = (byte)spec.Channels,
                 silence = spec.Silence,
                 samples = spec.Samples,
                 padding = spec.Padding,
